+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.2]
+- Updated `README.md` and documentation
+- `toLocal()` and `toUtc()` added to `general_datetime_interface`
+- `toLocal()` and `toUtc()` implemented for `jalali_datetime`
+
 ## [0.1.1]
 - Updated `README.md` and documentation
 - Jalali Datetime completed (Persian calendar)